requirements.txt
boto3
streamlit
argparse
datetime
streamlit_pdf_reader
pyyaml